+++ rpackage/sentiment.ai/R/choose_model.R
@@ -12,18 +12,21 @@ choose_model <- function(model){
 
   # model defaults/names
   # !! default_models is declared in constants.R !!
-  model_names    <- names(default_models)
+  model_names    <- c(names(default_models), names(openai_models))
 
   default_models <- setNames(object = "https://tfhub.dev/google/" %+% default_models,
-                             nm     = model_names)
+                             nm     = names(default_models))
+
+  all_models <- c(default_models, openai_models)
 
   # check whether model is IN model defaults
   if(model %in% model_names){
-    model <- default_models[model]
-  } else if(model %ni% default_models){
+    model <- all_models[model]
+  } else if(model %ni% all_models){
     warning(model, " model is not one of the built-in models: ",
             paste(model_names, collapse = ", "), ". ",
             "Overriding the defaults is allowed, but may or may not work! ",
+            "And TFhub model can be called with the full path, e.g. https://tfhub.dev/google/universal-sentence-encoder-large/5",
             "You're on your own from here on, cowboy! Godspeed!",
             call. = FALSE)
   } # END ifelse STATEMENT